publisher/requirements.txt
paho-mqtt
pyyaml
pyaml-env